Convert rating to int

+++ server/scraper/appstore_textpreprocess.py
@@ -62,5 +62,5 @@ def preprocess_appstore_data(df):
     df['title'] = df['title'].apply(preproccessing)
     df['bank'] = df['bank'].str.upper()
     df['source'] = "appstore"
-    
+    df['rating'] = df['rating'].astype(int)
     return df
